+++ src/components/sdk/MeetingControls/AudioOutputControl.tsx
@@ -2,12 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { DefaultBrowserBehavior } from 'amazon-chime-sdk-js';
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 
 import useSelectAudioOutputDevice from '../../../hooks/sdk/useSelectAudioOutputDevice';
 import { useAudioOutputs } from '../../../providers/DevicesProvider';
 import { useLocalAudioOutput } from '../../../providers/LocalAudioOutputProvider';
-import { DeviceType } from '../../../types';
 import { isOptionActive } from '../../../utils/device-utils';
 import { ControlBarButton } from '../../ui/ControlBar/ControlBarButton';
 import { Sound } from '../../ui/icons';
@@ -26,28 +25,40 @@ const AudioOutputControl: React.FC<Props> = ({
   const selectAudioOutput = useSelectAudioOutputDevice();
   const { devices, selectedDevice } = useAudioOutputs();
   const { isAudioOn, toggleAudio } = useLocalAudioOutput();
+  const [dropdownOptions, setDropdownOptions] = useState<PopOverItemProps[]>(
+    []
+  );
 
-  const handleClick = async (deviceId: string): Promise<void> => {
-    try {
-      if (new DefaultBrowserBehavior().supportsSetSinkId()) {
-        await selectAudioOutput(deviceId);
-      } else {
+  useEffect(() => {
+    const handleClick = async (deviceId: string): Promise<void> => {
+      try {
+        if (new DefaultBrowserBehavior().supportsSetSinkId()) {
+          await selectAudioOutput(deviceId);
+        } else {
+          console.error(
+            'AudioOutputControl cannot select audio output device because browser does not support setSinkId operation.'
+          );
+        }
+      } catch (error) {
         console.error(
-          'AudioOutputControl cannot select audio output device because browser does not support setSinkId operation.'
+          'AudioOutputControl failed to select audio output device'
         );
       }
-    } catch (error) {
-      console.error('AudioOutputControl failed to select audio output device');
-    }
-  };
+    };
 
-  const dropdownOptions: PopOverItemProps[] = devices.map(
-    (device: DeviceType) => ({
-      children: <span>{device.label}</span>,
-      checked: isOptionActive(selectedDevice, device.deviceId),
-      onClick: async () => await handleClick(device.deviceId),
-    })
-  );
+    const getDropdownOptions = async (): Promise<void> => {
+      const dropdownOptions = await Promise.all(
+        devices.map(async (device) => ({
+          children: <span>{device.label}</span>,
+          checked: await isOptionActive(selectedDevice, device.deviceId),
+          onClick: async () => await handleClick(device.deviceId),
+        }))
+      );
+      setDropdownOptions(dropdownOptions);
+    };
+
+    getDropdownOptions();
+  }, [devices, selectedDevice]);
 
   return (
     <>