+++ src/constants/index.ts
@@ -11,6 +11,6 @@ export const KEY_CODES = {
 
 export const VIDEO_INPUT_QUALITY = {
   '360p': '360p (nHD) @ 15 fps (600 Kbps max)',
-  '540p': '540p (qHD) @ 15 fps (1.0 Mbps max)',
+  '540p': '540p (qHD) @ 15 fps (1.4 Mbps max)',
   '720p': '720p (HD) @ 15 fps (1.4 Mbps max)',
 };