+++ src/hooks/sdk/useSelectVideoInputDevice.tsx
@@ -1,18 +1,18 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { Device, VideoTransformDevice } from 'amazon-chime-sdk-js';
+import { VideoInputDevice } from 'amazon-chime-sdk-js';
 import { useCallback } from 'react';
 
 import { useMeetingManager } from '../../providers/MeetingProvider';
 
 export const useSelectVideoInputDevice = (): ((
-  device: Device | VideoTransformDevice
+  device: VideoInputDevice
 ) => Promise<void>) => {
   const meetingManager = useMeetingManager();
 
-  const selectVideo = useCallback(async (deviceId: string) => {
-    await meetingManager.selectVideoInputDevice(deviceId);
+  const selectVideo = useCallback(async (device: VideoInputDevice) => {
+    await meetingManager.selectVideoInputDevice(device);
   }, []);
 
   return selectVideo;