+++ src/utils/device-utils.ts
@@ -1,11 +1,52 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export const isOptionActive = (
-  meetingManagerDeviceId: string | null,
+import {
+  AudioTransformDevice,
+  DefaultDeviceController,
+  Device,
+  isAudioTransformDevice,
+  isVideoTransformDevice,
+  VideoTransformDevice,
+} from 'amazon-chime-sdk-js';
+
+export const isOptionActive = async (
+  selectedDevice:
+    | Device
+    | AudioTransformDevice
+    | VideoTransformDevice
+    | null
+    | undefined,
   currentDeviceId: string
-): boolean => {
-  return currentDeviceId === meetingManagerDeviceId;
+): Promise<boolean> => {
+  const selectedDeviceId = await getDeviceId(selectedDevice);
+  return selectedDeviceId === currentDeviceId;
+};
+
+export const getDeviceId = async (
+  device:
+    | Device
+    | AudioTransformDevice
+    | VideoTransformDevice
+    | null
+    | undefined
+): Promise<string> => {
+  if (!device) {
+    return '';
+  }
+
+  let intrinsicDevice: Device | null;
+
+  if (isAudioTransformDevice(device) || isVideoTransformDevice(device)) {
+    intrinsicDevice = await device.intrinsicDevice();
+  } else {
+    intrinsicDevice = device;
+  }
+  const deviceId = DefaultDeviceController.getIntrinsicDeviceId(
+    intrinsicDevice
+  ) as string;
+
+  return deviceId;
 };
 
 export function isPrevNextUndefined<T>(prev: T, next: T): boolean {